* commit '08049329971b7fa67d86d12e1d87a5fa51de092c': (65 commits)
  More scaladoc improvements. I think this is the last set.
  Change GravitationalForce to ProportionalConstrainingForce - it isn't gravitational at all, and calling it such is mathematically confusing.
  Stop passing the number of nodes and edges into force application - nothing was using it anymore.
  Better docs on LayoutNodes and LayoutEdges, and removal of some extraneous parameters I found writing the better docs.
  Yet more scaladoc changes
  Better scaladocs for force-directed layout parameters.
  Better comment in scaling function.
  Include copyright notices in median calculator files.
  scaladoc comments for gravity
  Add scaladocs to repulsion forces. Sadly, this introduces more questions than it answers - I think there may be some basic problems in the math herein.
  Make the node overlap repulsion increase a constant - it was dependent on node size, which makes little sense (why would the proportional increase in repulsion between overlapping nodes in a smaller community be greater than those in a larger one?)  And in its old form, it would be counterproductive in any community whose radius was > 1000.0
  More scaladocs
  enables scalastyle plugin
  Turn off useProjects
  Style change, and scaladoc
  Some more commenting improvements, and removal of more unused parameters
  Change to parameters of ForceDirectedLayout.run implemented in visualizer.
  Some further scaladoc improvements - and the removal of a parameter I realized (when trying to document it) wasn't being used
  Some scaladoc improvements
  New parameter in layout config I forgot to add to test.
  ...

+++ src/scripts/layout.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
 
+
+
+# Include any needed common scripts
+SOURCE_LOCATION=$( cd $( dirname ${BASH_SOURCE[0]} ) && pwd )
+. ${SOURCE_LOCATION}/arg-parser.sh
+
+
+
 MAIN_JAR=../xdata-graph-0.1-SNAPSHOT/lib/xdata-graph.jar
 MAIN_CLASS=software.uncharted.graphing.layout.ClusteredGraphLayoutApp
-BASE_LOCATION=/user/nkronenfeld/graphing/timing
+BASE_LOCATION=/user/${USER}/graphs
 
 
 
 DATASET=
+REMOVE_EXISTING=false
 
 while [ "$1" != "" ]; do
 	case $1 in 
@@ -14,6 +23,8 @@ while [ "$1" != "" ]; do
 			shift
 			DATASET=$1
 			;;
+		-r | --refresh )
+			REMOVE_EXISTING=true
 	esac
 	shift
 done
@@ -22,6 +33,13 @@ if [ "${DATASET}" == "" ]; then
 	echo No dataset specified
 	exit
 fi
+
+
+
+# copy our config files into the dataset, if they're not already there
+CONFIG_COPIED=$(checkConfigFile ${SOURCE_LOCATION}/config/default-layout.conf ${DATASET}/layout.conf)
+
+
 
 pushd ${DATASET}
 
@@ -50,43 +68,95 @@ echo EXECUTORS: ${EXECUTORS} >> layout.log
 
 TIMEA=$(date +%s)
 
-echo Pushing to HDFS
-hdfs dfs -rm -r -skipTrash ${BASE_LOCATION}/${DATASET}
-hdfs dfs -mkdir ${BASE_LOCATION}/${DATASET}
-hdfs dfs -mkdir ${BASE_LOCATION}/${DATASET}/clusters
-hdfs dfs -put level_* ${BASE_LOCATION}/${DATASET}/clusters
-hdfs dfs -rm ${BASE_LOCATION}/${DATASET}/clusters/level_*/stats
+# Count the number of files that match the given name at the given location in HDFS
+function countHDFSFiles {
+	LOCATION=$1
+	RESULTS=`hdfs dfs -ls -d ${LOCATION} | wc | awk '{print $1}'`
+	if [ "" == "${RESULTS}" ]; then
+		echo 0
+	else
+		echo ${RESULTS}
+	fi
+}
+
+echo Checking base location
+if [ 0 -eq "$(countHDFSFiles ${BASE_LOCATION})" ]; then
+	echo Base location ${BASE_LOCATION} does not exists!
+	exit
+fi
+
+echo Checking dataset
+if [ "${DATASET}" == "" ]; then
+	echo No dataset specified
+	exit
+fi
+
+echo Checking file existence
+COPY_LOCAL=0
+if [ 0 -eq "$(countHDFSFiles ${BASE_LOCATION}/${DATASET})" ]; then
+	hdfs dfs -mkdir ${BASE_LOCATION}/${DATASET}
+	COPY_LOCAL=1
+else
+	LOCAL_CLUSTERS_EXIST=($(countHDFSFiles file:${PWD}/level_*))
+	CLUSTERS_EXIST=($(countHDFSFiles ${BASE_LOCATION}/${DATASET}/clusters))
+	LAYOUT_EXISTS=($(countHDFSFiles ${BASE_LOCATION}/${DATASET}/layout))
+
+	if [ "${REMOVE_EXISTING}" == "true" -a 1 -eq "${CLUSTERS_EXIST}" -a 0 -lt "${LOCAL_CLUSTERS_EXIST}" ]; then
+		hdfs dfs -rm -r ${BASE_LOCATION}/${DATASET}/clusters
+		COPY_LOCAL=1
+	elif [ 0 -eq "${CLUSTERS_EXIST}" -a 0 -lt "${LOCAL_CLUSTERS_EXIST}" ]; then
+		COPY_LOCAL=1
+	else
+		COPY_LOCAL=0
+	fi
+
+	if [ 1 -eq "${LAYOUT_EXISTS}" ]; then
+		hdfs dfs -rm -r ${BASE_LOCATION}/${DATASET}/layout
+	fi
+fi
+
+if [ 1 -eq "${COPY_LOCAL}" ]; then
+	echo Pushing to HDFS
+	hdfs dfs -mkdir ${BASE_LOCATION}/${DATASET}/clusters
+	hdfs dfs -put level_* ${BASE_LOCATION}/${DATASET}/clusters
+	hdfs dfs -rm ${BASE_LOCATION}/${DATASET}/clusters/level_*/stats
+fi
 
 TIMEB=$(date +%s)
 
 echo Starting layout run
+
+# These variables need to be exported for the config file
+export BASE_LOCATION
+export DATASET
+export MAX_LEVEL
+export PARTS
 
 echo spark-submit \
 	--class ${MAIN_CLASS} \
 	--num-executors ${EXECUTORS} \
 	--executor-cores 4 \
 	--executor-memory 10g \
+	--master yarn \
+	--deploy-mode client \
 	${MAIN_JAR} \
-	-source ${BASE_LOCATION}/${DATASET}/clusters \
-	-output ${BASE_LOCATION}/${DATASET}/layout \
-	-maxLevel ${MAX_LEVEL} \
-	-spark yarn-client \
-	-parts ${PARTS} >> layout.log
+	debug layout.conf >> layout.log
 
 spark-submit \
 	--class ${MAIN_CLASS} \
 	--num-executors ${EXECUTORS} \
 	--executor-cores 4 \
 	--executor-memory 10g \
+	--master yarn \
+	--deploy-mode client \
 	${MAIN_JAR} \
-	-source ${BASE_LOCATION}/${DATASET}/clusters \
-	-output ${BASE_LOCATION}/${DATASET}/layout \
-	-maxLevel ${MAX_LEVEL} \
-	-parts ${PARTS} |& tee -a layout.log
+	debug layout.conf |& tee -a layout.log
 
 # Note: Took out -spark yarn-client.  Should be irrelevant, but noted just in case I'm wrong.
 
 TIMEC=$(date +%s)
+
+cleanupConfigFile ${CONFIG_COPIED} layout.conf
 
 echo >> layout.log
 echo >> layout.log
@@ -100,4 +170,3 @@ echo Time to upload to HDFS: $(( ${TIMEB} - ${TIMEA} )) seconds
 echo Elapsed time for layout: $(( ${TIMEC} - ${TIMEB} )) seconds
 
 popd
-
